Move hex decoding functions into new hex.hpp/cpp
