+++ src/hex.cpp
@@ -9,7 +9,9 @@
 
 #include "hex.hpp"
 
+#include <array>
 #include <cassert>
+#include <stdexcept>
 
 namespace util {
 
@@ -32,6 +34,45 @@ std::string encode_hex(std::string const &in)
     result.reserve(in.size() * 2);
     encode_hex(in, &result);
     return result;
+}
+
+namespace {
+
+constexpr std::array<char, 256> const HEX_TABLE = {
+    0, 0, 0, 0,   0, 0, 0, 0,   0, 0, 0, 0,   0, 0, 0, 0,
+    0, 0, 0, 0,   0, 0, 0, 0,   0, 0, 0, 0,   0, 0, 0, 0,
+    0, 0, 0, 0,   0, 0, 0, 0,   0, 0, 0, 0,   0, 0, 0, 0,
+    0, 1, 2, 3,   4, 5, 6, 7,   8, 9, 0, 0,   0, 0, 0, 0,
+
+    0, 10, 11, 12,   13, 14, 15, 0,   0, 0, 0, 0,   0, 0, 0, 0,
+    0,  0,  0,  0,    0,  0,  0, 0,   0, 0, 0, 0,   0, 0, 0, 0,
+    0, 10, 11, 12,   13, 14, 15, 0,   0, 0, 0, 0,   0, 0, 0, 0,
+};
+
+} // anonymous namespace
+
+unsigned char decode_hex_char(char c) noexcept
+{
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-constant-array-index)
+    return HEX_TABLE[static_cast<std::size_t>(static_cast<unsigned char>(c))];
+}
+
+std::string decode_hex(std::string_view hex_string)
+{
+    if (hex_string.size() % 2 != 0) {
+        throw std::runtime_error{"Invalid wkb: Not a valid hex string"};
+    }
+
+    std::string wkb;
+    wkb.reserve(hex_string.size() / 2);
+
+    // NOLINTNEXTLINE(llvm-qualified-auto, readability-qualified-auto)
+    for (auto hex = hex_string.begin(); hex != hex_string.end();) {
+        unsigned int const c = decode_hex_char(*hex++);
+        wkb += static_cast<char>((c << 4U) | decode_hex_char(*hex++));
+    }
+
+    return wkb;
 }
 
 } // namespace util
