+++ tests/test-wkb.cpp
@@ -189,46 +189,3 @@ TEST_CASE("wkb: geometrycollection", "[NoDB]")
 
 TEST_CASE("wkb: invalid", "[NoDB]") { REQUIRE_THROWS(ewkb_to_geom("INVALID")); }
 
-TEST_CASE("wkb hex decode of valid and invalid hex characters")
-{
-    REQUIRE(decode_hex_char('0') == 0);
-    REQUIRE(decode_hex_char('9') == 9);
-    REQUIRE(decode_hex_char('a') == 0x0a);
-    REQUIRE(decode_hex_char('f') == 0x0f);
-    REQUIRE(decode_hex_char('A') == 0x0a);
-    REQUIRE(decode_hex_char('F') == 0x0f);
-    REQUIRE(decode_hex_char('#') == 0);
-    REQUIRE(decode_hex_char('@') == 0);
-    REQUIRE(decode_hex_char('g') == 0);
-    REQUIRE(decode_hex_char('G') == 0);
-    REQUIRE(decode_hex_char(0x7f) == 0);
-}
-
-TEST_CASE("wkb hex decode of valid hex string")
-{
-    std::string const hex{"0001020F1099FF"};
-    std::string const data = {0x00,
-                              0x01,
-                              0x02,
-                              0x0f,
-                              0x10,
-                              static_cast<char>(0x99),
-                              static_cast<char>(0xff)};
-
-    auto const result = decode_hex(hex);
-    REQUIRE(result.size() == hex.size() / 2);
-    REQUIRE(result == data);
-}
-
-TEST_CASE("wkb hex decode of empty string is okay")
-{
-    std::string const hex{};
-    REQUIRE(decode_hex(hex).empty());
-}
-
-TEST_CASE("wkb hex decode of string with odd number of characters fails")
-{
-    REQUIRE_THROWS(decode_hex("a"));
-    REQUIRE_THROWS(decode_hex("abc"));
-    REQUIRE_THROWS(decode_hex("00000"));
-}
